Added Supplementary Files
Ah, much better! All of the preliminary files are here! From here on out I'll be working on making those custom UI elements work with the plain-old-default UI elements that I usually use.

On that note, I've created the WelcomeUI and the ImageSetPickerUI, so now all the custom UI design should be finished. Now I just need to make it work!

I've also added a config file (for storing user config settings) and a testFile.txt (for testing read/write perms in directories).