Added Supplementary Files
I forgot to add the build folder and the dist folder, so those are now available if need be. The dist folder contains a VERY preliminary version of the program, so check that out if there is any interest for it. I may also need to update the README.txt file in the dist folder, but it overall has good content for the time being.