Updated Custom UI Files
The ImageSetPickerUI is complete! You can now completely initialize/reset the config file as needed. Now I just need to finish the WelcomeUI with all of the FileIO operations.

Updated README to reflect current project status. I also removed the "Config Object" and "Config Service" ideas because I don't think it will end up being necessary to use objects in this specific case. If I were using multiple configs, then that would be helpful, but that is not the case here.