Updated Custom UI Files
Almost finished with the WelcomeUI file, with some specific exceptions. Those exceptions include setting up source directory (if not set up) before and after transfer, writing to the config, and working on "Smart Sources". I will also need to do lots of testing to ensure that this will do the job properly!