Updated Custom UI Files
Welp, I decided to "overhaul" the WelcomeUI a little bit. In doing so, I made a little bit more work for myself, but I think in the end it will make it look much more polished. That "little bit of work" is adding in a SettingsUI.

The Settings UI will handle... well... the settings for all of the directories and the number of image sets that will need to be made. That gives the user a little more control if absolutely necessary. Also, the "Reset Config" option was moved to the SettingsUI as well.