Updated Custom UI Files
Well, after some long and difficult work, I decided to take a simpler approach with the Smart Source Picker. Instead of trying to be some sort of "AI" (it's not...), I may as well just try to make intelligent guesses for which folder is the next source. For example, if months/years are being utilized in prior folder names, then the next month/year will be suggested.

On a side note, I also adjusted how image set folders are created in the main process.