Updated Custom UI Files
After some mild testing, I have identified a handful of issues with the Settings UI, which I can simply just fix with a second constructor. That will be in the works, but for now, here's this!

I also updated the StatsUI so it can read a brand new config without breaking... Oops!