Updated Custom UI Files
Added some functionality to SettingsUI in case the image set number changes. Image Set Folders will be added or removed based on the new number chosen by the user. Of course, if the folder has images or files present in one of these folders set for deletion, then it will be left alone.

As for WelcomeUI, I updated some code that fixed weird bugs, like accidentally creating a "0 Print" instead of "1 Print"... Whoops.