Updated Custom UI Files
Updated the StatsUI data collection to more properly... collect data!

I was also really impressed with the operation of the Bring Back Photos function in the WelcomeUI! On its first test it worked properly! I'm a little embarrassed to not have tested it earlier, but it would have left incomplete files all over the place. It would have been more work to clean up than it was worth. I'm now working on updating the smartSourcePicker to detect consistent duplicates and pick that source. Huzzah!